Move setenv/unsetenv to azcore and remove aztest dependency from vulkan back end
