+++ Code/Framework/AzCore/Platform/Common/WinAPI/AzCore/Utils/Utils_WinAPI.cpp
@@ -73,6 +73,14 @@ namespace AZ
             return result != nullptr;
         }
 
+        bool SetEnv(const char* envname, const char* envvalue, [[maybe_unused]] bool overwrite)
+        {
+            return _putenv_s(envname, envvalue);
+        }
 
+        bool UnsetEnv(const char* envname)
+        {
+            return SetEnv(envname, "", 1);
+        }
     }
 }
