+++ Code/Framework/AzTest/AzTest/Platform/Common/Unimplemented/AzTest/Utils_Unimplemented.cpp
@@ -12,14 +12,6 @@ namespace AZ
 {
     namespace Test
     {
-        bool SetEnv([[maybe_unused]] const char* envname, [[maybe_unused]] const char* envvalue, [[maybe_unused]] bool overwrite)
-        {
-            return false;
-        }
-
-        bool UnsetEnv([[maybe_unused]] const char* envname)
-        {
-            return false;
-        }
+        //Unimplemented methods will go here
     } // namespace Test
 } // namespace AZ
